MVP restructuring
