+++ app/src/main/java/com/example/photogallery/Presenter/IMainPresenter.java
@@ -2,8 +2,6 @@ package com.example.photogallery.Presenter;
 
 import android.content.Intent;
 
-import com.example.photogallery.ImageAdapter;
-
 public interface IMainPresenter {
     void handleRequestImageCapture(ImageAdapter imageAdapter);
     void handlePhotoPreviewUpdates(Intent data);
