+++ app/src/main/java/com/example/photogallery/Presenter/ISearchPresenter.java
@@ -1,0 +1,4 @@
+package com.example.photogallery.Presenter;
+
+public interface ISearchPresenter {
+}
