+++ app/src/main/java/com/example/photogallery/Presenter/PhotoPreviewPresenter.java
@@ -1,0 +1,4 @@
+package com.example.photogallery.Presenter;
+
+public class PhotoPreviewPresenter {
+}
