+++ app/src/main/java/com/example/photogallery/View/MainView.java
@@ -1,7 +1,13 @@
 package com.example.photogallery.View;
 
-import com.example.photogallery.ImageAdapter;
+import android.view.View;
+
+import com.example.photogallery.Presenter.ImageAdapter;
+
+import java.io.IOException;
 
 public interface MainView {
     void updateGridView(ImageAdapter imageAdapter);
+    void openSearchActivity(View v);
+    void takePhoto(View v) throws IOException;
 }
