+++ app/src/main/java/com/example/photogallery/View/PhotoPreviewView.java
@@ -1,0 +1,4 @@
+package com.example.photogallery.View;
+
+public interface PhotoPreviewView {
+}
