+++ app/src/main/java/com/example/photogallery/View/SearchView.java
@@ -1,0 +1,4 @@
+package com.example.photogallery.View;
+
+public interface SearchView {
+}
